requirements.txt
django
djangorestframework
environs
psychopg2-binary
djangorestframework-simplejwt
python-dotenv
